dev/4_microservices/random/random-front/requirements.txt
flask ~= 1.1
Jinja2 ~= 2.11
pytest ~= 5.4
markupsafe==2.0.1

grpc-interceptor >= 0.11.0
grpcio-tools >= 1.30